[Refactor] Move lru to base
