+++ be/src/exprs/ai_functions.h
@@ -14,9 +14,9 @@
 
 #pragma once
 
+#include "base/container/lru_cache.h"
 #include "exprs/builtin_functions.h"
 #include "exprs/function_helper.h"
-#include "util/lru_cache.h"
 
 namespace starrocks {
