+++ be/src/fs/fd_cache.cpp
@@ -16,7 +16,7 @@
 
 #include <unistd.h>
 
-#include "util/lru_cache.h"
+#include "base/container/lru_cache.h"
 
 namespace starrocks {
