+++ be/src/runtime/load_channel.cpp
@@ -36,6 +36,7 @@
 
 #include <memory>
 
+#include "base/container/lru_cache.h"
 #include "base/string/faststring.h"
 #include "common/runtime_profile.h"
 #include "common/tracer.h"
@@ -50,7 +51,6 @@
 #include "runtime/mem_tracker.h"
 #include "runtime/starrocks_metrics.h"
 #include "util/compression/block_compression.h"
-#include "util/lru_cache.h"
 
 #define RETURN_RESPONSE_IF_ERROR(stmt, response)                                      \
     do {                                                                              \
