+++ be/src/storage/sstable/table.cpp
@@ -7,6 +7,7 @@
 #include <butil/time.h> // NOLINT
 
 #include "base/coding.h"
+#include "base/container/lru_cache.h"
 #include "base/debug/trace.h"
 #include "common/status.h"
 #include "fs/fs.h"
@@ -19,7 +20,6 @@
 #include "storage/sstable/format.h"
 #include "storage/sstable/options.h"
 #include "storage/sstable/two_level_iterator.h"
-#include "util/lru_cache.h"
 
 namespace starrocks::sstable {
