+++ be/src/storage/txn_manager.h
@@ -53,6 +53,7 @@
 
 #include "agent/status.h"
 #include "base/concurrency/countdown_latch.h"
+#include "base/container/lru_cache.h"
 #include "base/time/time.h"
 #include "common/status.h"
 #include "common/thread/threadpool.h"
@@ -66,7 +67,6 @@
 #include "storage/rowset/rowset.h"
 #include "storage/rowset/rowset_meta.h"
 #include "storage/tablet.h"
-#include "util/lru_cache.h"
 
 namespace starrocks {
