+++ be/src/util/CMakeLists.txt
@@ -42,7 +42,6 @@ set(UTIL_FILES
 # coding_util.cpp
   thrift_rpc_helper.cpp
   frame_of_reference_coding.cpp
-  lru_cache.cpp
   debug/query_trace_impl.cpp
   memory_lock.cpp
   internal_service_recoverable_stub.cpp
