+++ be/src/util/llm_cache.cpp
@@ -14,8 +14,8 @@
 
 #include "llm_cache.h"
 
+#include "base/container/lru_cache.h"
 #include "base/crypto/md5.h"
-#include "util/lru_cache.h"
 
 namespace starrocks {
